Adding static_cast to uint8_t in order to limit the maximum type size to uint8_t and prevent automatic upcasts to higher types if 'bit_pos' exceeds 7

+++ include/spectre/util/BitManip.h
@@ -41,17 +41,17 @@ namespace util
 
 inline void setBit(volatile uint8_t * reg, uint8_t const bit_pos)
 {
-  *reg |= (1<<bit_pos);
+  *reg |= static_cast<uint8_t>(1<<bit_pos);
 }
 
 inline void clrBit(volatile uint8_t * reg, uint8_t const bit_pos)
 {
-  *reg &= ~(1<<bit_pos);
+  *reg &= ~(static_cast<uint8_t>(1<<bit_pos));
 }
 
 inline void toogleBit(volatile uint8_t * reg, uint8_t const bit_pos)
 {
-  *reg ^= (1<<bit_pos);
+  *reg ^= static_cast<uint8_t>(1<<bit_pos);
 }
 
 /**************************************************************************************
